docs(examples): add better documentation on directive persistence used in the example
