+++ examples/react-native/src/utils/persistence/index.ts
