+++ examples/react-native/src/utils/persistence/persistLink.ts
