+++ examples/react-native/src/utils/persistence/transform.ts
